Altered pk's from ints to longs
git-svn-id: https://svn.mothershipdev.com:9443/svn/PaymentSense_MothershipV2_Core/branches/Usermanagement@182 0f896bb5-3ab0-2d4e-82a7-85ef10020915

+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/GroupRole.cs
@@ -8,7 +8,7 @@ namespace PS.Mothership.Core.Common.Dto
 {
     public class GroupRole
     {
-        public int GroupId { get; set; }
-        public int RoleId { get; set; }
+        public long GroupId { get; set; }
+        public long RoleId { get; set; }
     }
 }
